pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thumbcraft"
version = "0.1.0"
description = "Production-grade YouTube thumbnail generator CLI powered by Gemini 3 Pro Image"
requires-python = ">=3.11"
authors = [{ name = "ThumbCraft" }]
readme = "README.md"

dependencies = [
    "typer[all]>=0.12.0",
    "google-genai>=1.0.0",
    "google-cloud-aiplatform>=1.60.0",
    "google-api-python-client>=2.100.0",
    "Pillow>=10.0.0",
    "PyYAML>=6.0",
    "Jinja2>=3.1.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "sqlite-utils>=3.35",
    "feedparser>=6.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
]

[project.scripts]
thumbcraft = "cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["cli*", "core*", "generators*", "compositing*", "preview*", "scraper*", "agents*", "studio*"]
exclude = ["cache*", "output*", "data*", "channels*", "prompts*"]
